chore: merge resource hints plugins
