+++ packages/core/src/provider/provider.ts
@@ -76,8 +76,7 @@ export const rspackProvider: RsbuildProvider = async ({
         plugins.inlineChunk(),
         plugins.bundleAnalyzer(),
         plugins.rsdoctor(),
-        plugins.networkPerformance(),
-        plugins.preloadOrPrefetch(),
+        plugins.resourceHints(),
         plugins.performance(),
         plugins.server(),
         plugins.moduleFederation(),
